Рефакторинг моделей
